tools/requirements.txt
boto3==1.28.14
botocore==1.31.14
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.1
google-auth==2.22.0
idna==3.4
jmespath==1.0.1
kubernetes==27.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.26.0
oauthlib==3.2.2
pandas==2.1.1
pandas-stubs==2.1.1.230928
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
s3transfer==0.6.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-toml==0.10.8.7
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.2
